LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-blob-with-no-mime-type-expected.txt


FAIL Synchronous blob loading with no mime type assert_equals: expected "NO" but got ""
FAIL Asynchronous blob loading with no mime type assert_equals: expected "NO" but got ""
FAIL Synchronous blob loading with invalid mime type assert_equals: expected "NO" but got ""
FAIL Asynchronous blob loading with invalid mime type assert_equals: expected "NO" but got ""
